First commit to restructure the project. A) Tasks were separated into indivual files b) The extension file was moved to its own folder named "Extension".
