+++ src/main/groovy/net/wasdev/wlp/gradle/plugins/tasks/LibertyRunTask.groovy
@@ -1,0 +1,39 @@
+/**
+ * (C) Copyright IBM Corporation 2015.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package net.wasdev.wlp.gradle.plugins.tasks;
+
+import net.wasdev.wlp.gradle.plugins.tasks.AbstractTask.LibertyListener
+import org.gradle.api.tasks.TaskAction
+
+import com.ibm.wsspi.kernel.embeddable.Server
+import com.ibm.wsspi.kernel.embeddable.ServerBuilder
+import com.ibm.wsspi.kernel.embeddable.Server.Result
+import com.ibm.wsspi.kernel.embeddable.ServerEventListener
+import com.ibm.wsspi.kernel.embeddable.ServerEventListener.ServerEvent
+import com.ibm.wsspi.kernel.embeddable.ServerEventListener.ServerEvent.Type
+
+class LibertyRunTask extends AbstractTask {
+    
+    @TaskAction
+    void printResponse() {    
+        ServerBuilder builder = getServerBuilder(project);
+        LibertyListener listener = new LibertyListener()
+        builder.setServerEventListener(listener)
+        Result result = builder.build().start().get()
+        if (!result.successful()) throw result.getException()
+        while (!Type.STOPPED.equals(listener.next().getType())) {}
+    }
+}
